fix: implement workaround for multiple map components target stealing
using the same map state to render multiple map components is not supported. with this fix when a
new map component is rendered it will steal the target from the existing map component. on destroy
the previous map component will automatically reinitialized itself. this helps in situation when an
application temporary create a new map instance (e.g. in a modal or a foreground window)